remove key and always just use uri
key had been devised at first to map a uri to a single file, but we needed to number them anyway to make the derived cache file names unique (a cache file can be evicted and created again, should be a different file). So key was useless and gets removed hereby.